requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
more-itertools==4.1.0
parsel==1.4.0
pdfminer3k==1.3.1
pluggy==0.6.0
ply==3.11
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyMySQL==0.8.1
pyOpenSSL==17.5.0
pytest==3.5.1
queuelib==1.5.0
requests==2.18.4
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
Twisted==18.4.0
urllib3==1.22
w3lib==1.19.0
webencodings==0.5.1
zope.interface==4.5.0
